google_api/requirements.txt
fastapi==0.63.0
google-api-core==1.26.1
google-api-python-client==2.0.2
uvicorn==0.13.4
python-dateutil==2.8.1
